refactor: move util function to util file
