+++ src/components/ImageUpload.tsx
@@ -10,18 +10,6 @@ type Props = {
 const allowedMimes = ['image/png', 'image/jpeg'];
 const allowedExts = ['.png', '.jpg', '.jpeg'];
 
-export function toBase64(file: File): Promise<string> {
-  return new Promise((resolve, reject) => {
-    const reader = new FileReader();
-    reader.onerror = () => reject(new Error('Failed to read file'));
-    reader.onload = () => {
-      const result = reader.result as string;
-      resolve(result.split(',')[1] || '');
-    };
-    reader.readAsDataURL(file);
-  });
-}
-
 export default function ImageUpload({
   name = 'avatar',
   maxSizeMB = 2,
